requirements.txt
langchain
langchain_openai
openai
python-dotenv
streamlit
streamlit_mic_recorder
watchdog